return error from engine factory func
